scraper-service/requirements.txt
playwright==1.42.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
dnspython==2.6.1
